fix: マイグレーション関数の戻り値をINTEGERからBIGINTに変更
SUM(integer)はPostgreSQLでBIGINTを返すため、既存の
get_top_users_posting_countと同様にBIGINTに統一し、
オーバーフローの可能性を排除。

+++ supabase/migrations/20260131193020_add_get_total_posting_count_by_mission.sql
@@ -5,14 +5,14 @@ CREATE OR REPLACE FUNCTION get_total_posting_count_by_mission(
   target_mission_id UUID,
   p_season_id UUID DEFAULT NULL
 )
-RETURNS INTEGER
+RETURNS BIGINT
 LANGUAGE plpgsql
 STABLE
 SECURITY DEFINER
 SET search_path = public, pg_temp
 AS $$
 DECLARE
-  total_count INTEGER := 0;
+  total_count BIGINT := 0;
 BEGIN
   IF p_season_id IS NOT NULL THEN
     -- Season-based counting
